refac: add todo service file
